Add a send message to the mailer that accepts a swift message.
This should be used all times. The email object should be build outside
of the mailer and send trough the mailer. That will clean up
dependencies on the mailer.
