+++ src/Common/Mailer.php
@@ -9,10 +9,7 @@ namespace Common;
  * file that was distributed with this source code.
  */
 
-use \TijsVerkoyen\CssToInlineStyles\CssToInlineStyles;
 use Frontend\Core\Engine\Model;
-use Frontend\Core\Engine\Template;
-use Backend\Core\Engine\Template as BackendTemplate;
 
 /**
  * This class will send mails
@@ -27,6 +24,8 @@ use Backend\Core\Engine\Template as BackendTemplate;
 class Mailer
 {
     /**
+     * @deprecated build a message object and send it trough the send method
+     *
      * Adds an email to the queue.
      *
      * @param string $subject      The subject for the email.
@@ -75,148 +74,33 @@ class Mailer
         $replyToEmail = ($replyToEmail === null) ? (string) $replyTo['email'] : $replyToEmail;
         $replyToName = ($replyToName === null) ? (string) $replyTo['name'] : $replyToName;
 
-        // build html
-        $html = '';
-        if ($isRawHTML) {
-            $html = $template;
-        } else {
-            $html = $this->getTemplateContent($template, $variables);
-        }
-        $html = $this->relativeToAbsolute($html);
-        if ($addUTM === true) {
-            $html = $this->addUTM($html, $subject);
-        }
-
-        $transport = \Common\Mailer\Transport::newInstance();
-
-        $message = \Swift_message::newInstance()
-            ->setSubject($subject)
+        $message = \Common\Mailer\Message::newInstance($subject)
             ->setFrom(array($fromEmail => $fromName))
             ->setTo(array($toEmail => $toName))
             ->setReplyTo(array($replyToEmail => $replyToName))
-            ->setBody($html, 'text/html')
+            ->parseHtml($template, $variables, $isRawHTML, $addUTM)
+            ->setPlainText($plainText)
+            ->addAttachments($attachments);
         ;
 
-        if ($plainText !== null) {
-            $message->addPart($plainText, 'text/plain');
-        }
-
         // attachments added
-        if (!empty($attachments)) {
-            // add attachments one by one
-            foreach ($attachments as $attachment) {
-                // only add existing files
-                if (is_file($attachment)) {
-                    $message->attach(\Swift_Attachment::fromPath($attachment));
-                }
-            }
-        }
-
-        $mailer = \Swift_Mailer::newInstance($transport);
-        $mailer->send($message);
+        $this->send($message);
 
         // trigger event
         Model::triggerEvent('Core', 'after_email_sent', array('message' => $message));
     }
 
     /**
-     * @param string $html    The html to convert links in.
-     * @param string $subject The subject of the mail
-     * @return string
-     */
-    private function addUTM($html, $subject)
-    {
-        // match links
-        $matches = array();
-        preg_match_all('/href="(http:\/\/(.*))"/iU', $html, $matches);
-
-        // any links?
-        $utm = array('utm_source' => 'mail', 'utm_medium' => 'email', 'utm_campaign' => Uri::getUrl($subject));
-        if (isset($matches[0]) && !empty($matches[0])) {
-            $searchLinks = array();
-            $replaceLinks = array();
-
-            // loop old links
-            foreach ($matches[1] as $i => $link) {
-                $searchLinks[] = $matches[0][$i];
-                $replaceLinks[] = 'href="' . Model::addURLParameters($link, $utm) . '"';
-            }
-
-            $html = str_replace($searchLinks, $replaceLinks, $html);
-        }
-
-        return $html;
-    }
-
-    /**
-     * Returns the content from a given template
+     * send a Swift_Message like message
      *
-     * @param string $template  The template to use.
-     * @param array  $variables The variables to assign.
-     * @return string
+     * @param  Swift_mime_message $message
+     * @return boolean
      */
-    private function getTemplateContent($template, $variables = null)
+    public function send(\Swift_Mime_Message $message)
     {
-        // new template instance
-        $tpl = null;
-        if (APPLICATION === 'Backend') {
-            $tpl = new BackendTemplate(false);
-        } else {
-            $tpl = new Template(false);
-        }
+        $transport = \Common\Mailer\Transport::newInstance();
+        $mailer = \Swift_Mailer::newInstance($transport);
 
-        // set some options
-        $tpl->setForceCompile(true);
-
-        // variables were set
-        if (!empty($variables)) {
-            $tpl->assign($variables);
-        }
-
-        // grab the content
-        $content = $tpl->getContent($template);
-
-        // replace internal links/images
-        $search = array('href="/', 'src="/');
-        $replace = array('href="' . SITE_URL . '/', 'src="' . SITE_URL . '/');
-        $content = str_replace($search, $replace, $content);
-
-        // create instance
-        $cssToInlineStyles = new CSSToInlineStyles();
-
-        // set some properties
-        $cssToInlineStyles->setHTML($content);
-        $cssToInlineStyles->setUseInlineStylesBlock(true);
-        $cssToInlineStyles->setEncoding(SPOON_CHARSET);
-
-        // return the content
-        return (string) $cssToInlineStyles->convert();
-    }
-
-    /**
-     * @param string $html  The html to convert links in.
-     * @return string
-     */
-    private function relativeToAbsolute($html)
-    {
-        // get internal links
-        $matches = array();
-        preg_match_all('|href="/(.*)"|i', $html, $matches);
-
-        // any links?
-        if (!empty($matches[0])) {
-            $search = array();
-            $replace = array();
-
-            // loop the links
-            foreach ($matches[0] as $key => $link) {
-                $search[] = $link;
-                $replace[] = 'href="' . SITE_URL . '/' . $matches[1][$key] . '"';
-            }
-
-            $html = str_replace($search, $replace, $html);
-        }
-
-        return $html;
+        return $mailer->send($message);
     }
 }
